requirements.txt
playwright
google-auth
google-auth-oauthlib
google-api-python-client
requests
rapidfuzz
python-dotenv
anthropic
flask
